requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
